Add files via upload
These are the completed tasks for lab 3. They involve the use of a buzzer, button and potentiometer and introduce the developer to the "tone" method. The aim was to use the potentiometer to alter the frequency of the square wave outputted by the buzzer. 

The circuity includes an external pull up resistor to prevent the debouncing problem.
A challenge is to remove the external resistor and instead use the internal pull up resistors of the Arduino Uno.
A second challenge is to remove the external resistor and use the Serial Monitor to compare how the button values differ when the btnPin is set to INPUT and then INPUT_PULLUP.
A final challenge is to redo the entire project using INPUT_PULLUP